+++ RoutineSRC/CMazeSolutionStorage.h
@@ -7,13 +7,13 @@ class CMazeModel;
 class CMazeSolutionStorage
 {
 public:
-    explicit CMazeSolutionStorage(uint16_t expectedAmount = 0);
+   explicit CMazeSolutionStorage(uint16_t expectedAmount = 0);
 
-    const CMazeModel* getModelAt(const uint16_t index) const;
-    int getStorageSize() const;
-    void pushBackModel(const CMazeModel& model);
-    void clearStorage();
+   const CMazeModel* getModelAt(const uint32_t index) const;
+   uint32_t getStorageSize() const;
+   void pushBackModel(const CMazeModel& model);
+   void clearStorage();
 
 private:
-    std::vector<CMazeModel> mModelStorage;
+   std::vector<CMazeModel> mModelStorage;
 };
